Simplify the cycling on the rest of sprites
After all sprites have been properly put through the cycle, there is an
amount of sprites which are just leftovers and need to be reset to an
out of screen position. The loop that did this was unnecessarily complex
and we could just rely on the 'y' register wrapping around.

This has the benefit that is less error prone, even if there's still
some glitches which most probably come from by-one errors and such.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>