Decrease the capacity of bullets on screen
This fixes any frame drops I have experienced so far. This is not a real
fix since it is more likely that there's some part of the code that is
simply too slow. At some point I should investigate where is that, but
it will probably reappear whenever I start looking into enemies.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>