assets: Duplicate the UFO kind of sprite
At first I thought that I could save up some space and then mirror the
sprite horizontally, but it actually makes things easier to have the UFO
enemy in its entirety as it makes initialization and moving the object
easier. This is indeed possible because we have plenty of room left on
the CHR bank.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>